+++ providers/capz/standard/capz_test.go
@@ -14,13 +14,10 @@ type ClusterValues struct {
 }
 
 var _ = Describe("Common tests", func() {
-	common.Run(&common.TestConfig{
-		// Disabled until https://github.com/giantswarm/roadmap/issues/2693
-		AutoScalingSupported: false,
-		BastionSupported:     false,
-		TeleportSupported:    true,
-		// Disabled until wildcard ingress support is added
-		ExternalDnsSupported:         false,
-		ControlPlaneMetricsSupported: true,
-	})
+	cfg := common.NewTestConfigWithDefaults()
+	// Disabled until https://github.com/giantswarm/roadmap/issues/2693
+	cfg.AutoScalingSupported = false
+	// Disabled until wildcard ingress support is added
+	cfg.ExternalDnsSupported = false
+	common.Run(cfg)
 })