+++ providers/capz/upgrade/capz_test.go
@@ -11,13 +11,10 @@ var _ = Describe("Basic upgrade test", Ordered, func() {
 	upgrade.Run(upgrade.NewTestConfigWithDefaults())
 
 	// Finally run the common tests after upgrade is completed
-	common.Run(&common.TestConfig{
-		// Disabled until https://github.com/giantswarm/roadmap/issues/2693
-		AutoScalingSupported: false,
-		BastionSupported:     false,
-		TeleportSupported:    true,
-		// Disabled until wildcard ingress support is added
-		ExternalDnsSupported:         false,
-		ControlPlaneMetricsSupported: true,
-	})
+	cfg := common.NewTestConfigWithDefaults()
+	// Disabled until https://github.com/giantswarm/roadmap/issues/2693
+	cfg.AutoScalingSupported = false
+	// Disabled until wildcard ingress support is added
+	cfg.ExternalDnsSupported = false
+	common.Run(cfg)
 })