+++ providers/eks/standard/eks_test.go
@@ -7,11 +7,14 @@ import (
 )
 
 var _ = Describe("Common tests", func() {
-	common.Run(&common.TestConfig{
-		AutoScalingSupported: true,
-		BastionSupported:     false,
-		ExternalDnsSupported: true,
-		// EKS does not have metrics for k8s control plane components.
-		ControlPlaneMetricsSupported: false,
-	})
+	cfg := common.NewTestConfigWithDefaults()
+	// EKS does not have metrics for k8s control plane components.
+	cfg.ControlPlaneMetricsSupported = false
+	// EKS doesn't have any of the Giant Swarm apps deployed
+	cfg.ObservabilityBundleInstalled = false
+	cfg.SecurityBundleInstalled = false
+	cfg.ExternalDnsSupported = false
+	cfg.AutoScalingSupported = false
+	cfg.CertManagerSupported = false
+	common.Run(cfg)
 })